moving AionContractDetailsImpl, AionRepositoryCache and ContractDetailsCacheImpl to modAionImpl
