+++ modAion/src/module-info.java
@@ -9,7 +9,6 @@ module aion.zero {
     requires commons.lang3;
 
     exports org.aion.zero.api;
-    exports org.aion.zero.db;
     exports org.aion.zero.types;
     exports org.aion.zero.exceptions;
 }
