+++ modAionImpl/src/org/aion/zero/impl/db/AbstractRepositoryCache.java
@@ -1,4 +1,4 @@
-package org.aion.mcf.db;
+package org.aion.zero.impl.db;
 
 import static org.aion.crypto.HashUtil.h256;
 import static org.aion.util.bytes.ByteUtil.EMPTY_BYTE_ARRAY;
@@ -15,6 +15,7 @@ import org.aion.interfaces.db.RepositoryCache;
 import org.aion.log.AionLoggerFactory;
 import org.aion.log.LogEnum;
 import org.aion.mcf.core.AccountState;
+import org.aion.mcf.db.IBlockStoreBase;
 import org.aion.types.Address;
 import org.aion.types.ByteArrayWrapper;
 import org.slf4j.Logger;
