+++ modAionImpl/src/org/aion/zero/impl/db/AionRepositoryImpl.java
@@ -21,7 +21,6 @@ import org.aion.interfaces.db.RepositoryCache;
 import org.aion.interfaces.db.RepositoryConfig;
 import org.aion.mcf.core.AccountState;
 import org.aion.mcf.db.AbstractRepository;
-import org.aion.mcf.db.ContractDetailsCacheImpl;
 import org.aion.mcf.db.TransactionStore;
 import org.aion.mcf.ds.ObjectDataSource;
 import org.aion.mcf.trie.SecureTrie;
@@ -32,7 +31,6 @@ import org.aion.p2p.V1Constants;
 import org.aion.types.Address;
 import org.aion.types.ByteArrayWrapper;
 import org.aion.util.conversions.Hex;
-import org.aion.zero.db.AionRepositoryCache;
 import org.aion.zero.impl.config.CfgAion;
 import org.aion.zero.impl.sync.DatabaseType;
 import org.aion.zero.impl.types.AionBlock;
