+++ modAionImpl/src/org/aion/zero/impl/db/ContractDetailsAion.java
@@ -2,7 +2,6 @@ package org.aion.zero.impl.db;
 
 import org.aion.interfaces.db.ContractDetails;
 import org.aion.interfaces.db.DetailsProvider;
-import org.aion.zero.db.AionContractDetailsImpl;
 
 /**
  * Contract details provider for Aion.
