+++ modAionImpl/src/org/aion/zero/impl/db/ContractDetailsCacheImpl.java
@@ -1,4 +1,4 @@
-package org.aion.mcf.db;
+package org.aion.zero.impl.db;
 
 import java.util.Arrays;
 import java.util.HashMap;
@@ -7,6 +7,7 @@ import java.util.Map.Entry;
 import java.util.Objects;
 import org.aion.interfaces.db.ByteArrayKeyValueStore;
 import org.aion.interfaces.db.ContractDetails;
+import org.aion.mcf.db.AbstractContractDetails;
 import org.aion.types.Address;
 import org.aion.types.ByteArrayWrapper;
